@@ -1,43 +1,43 @@
-import React, {Component, Fragment} from 'react'
-import ReactDOM from 'react-dom'
+import React, {useState, useEffect, Fragment} from 'react'
 import Sampler from './sampler'
 import Keyboard from './keyboard'
 import './css/app.css'
 
 const keyVals = ['r', 't', 'y', 'u', 'f', 'g', 'h', 'j']
 
-class App extends Component {
-  state = {
-    pressedKey: ''
+const App = () => {
+  const [pressedKey, setPressedKey] = useState('')
+
+  const startSamplers = context => {
+    if(context.state === 'suspended') context.resume()
   }
 
-  handleChange = input => {
-    console.log("Input changed", input)
+  const handleKeyPress = ({key}) => {
+    setPressedKey(key)
   }
 
-  handleKeyPress = pressedKey => {
-    console.log('pressedKey:  ', pressedKey)
-    this.setState({pressedKey})
-  }
-
-  render() {
-    const {handleChange, handleKeyPress} = this
-    const {pressedKey} = this.state
-    return (
-      <Fragment>
-        {keyVals.map(keyVal => <Sampler
+  const row1 = keyVals.slice(0, 4)
+  const row2 = keyVals.slice(4)
+  return (
+    <div tabIndex='0'>
+      <div className='row'>
+        {row1.map(keyVal => <Sampler
           keyVal={keyVal}
-          pressedKey={pressedKey}
-          handleKeyPress={handleKeyPress}
+          start={startSamplers}
         />)}
-        <Keyboard
-          keyVals={keyVals}
-          handleChange={handleChange}
-          handleKeyPress={handleKeyPress}
-        />
-      </Fragment>
-    )
-  }
+      </div>
+      <div className='row'>
+        {row2.map(keyVal => <Sampler
+          keyVal={keyVal}
+          start={startSamplers}
+        />)}
+      </div>
+      <Keyboard
+        keyVals={keyVals}
+        handleKeyPress={handleKeyPress}
+      />
+    </div>
+  )
 }
 
 export default App
